switch to 32bit mono (does not work)

+++ CM4/Peripheral/Inc/pcm5102.h
@@ -26,8 +26,8 @@ extern "C" {
 
 /* Exported functions prototypes ---------------------------------------------*/
 void pcm5102_Init(void);
-int16_t pcm5102_GetAudioData(uint32_t index);
-volatile int16_t * pcm5102_GetDataPtr(uint32_t index);
+int32_t pcm5102_GetAudioData(uint32_t index);
+volatile int32_t * pcm5102_GetDataPtr(uint32_t index);
 
 /* Private defines -----------------------------------------------------------*/
